requirements.txt
aiofiles
aiogram
aiohappyeyeballs
aiohttp
aiosignal
amqp
annotated-types
anyio
attrs
billiard
cachetools
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
dataclasses-json
datadispatch
deepseek
distro
dnspython
duckduckgo_search
emoji
fastapi
feedparser
filelock
filetype
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h2
hpack
httpcore
httplib2
httpx
httpx-sse
huggingface-hub
hyperframe
idna
Jinja2
jiter
joblib
jsonpatch
jsonpointer
keybert
kombu
langchain
langchain-community
langchain-core
langchain-deepseek
langchain-google-genai
langchain-ollama
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
magic-filter
markdown-it-py
MarkupSafe
marshmallow
mdurl
motor
mpmath
multidict
mypy_extensions
networkx
ollama
openai
orjson
ormsgpack
packaging
pandas
pillow
primp
prompt_toolkit
propcache
proto-plus
protobuf
pyaes
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
Pygments
pymongo
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
qdrant-client
redis
regex
requests
requests-toolbelt
rich
rq
rsa
safetensors
sentence-transformers
sgmllib3k
six
sniffio
SQLAlchemy
starlette
sympy
tasks
Telethon
tenacity
threadpoolctl
tiktoken
tokenizers
tqdm
transformers
typing-inspect
typing-inspection
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
uvicorn
vine
wcwidth
xxhash
yarl
zstandard
